Added a way to load the names of txt files in the collection and a plot of one of these txt files.
The new feature allows users to load the names of all txt files in a directory and its subdirectories. The names are stored in a list that can be used for further processing.

In addition, a plot of one of the txt files is generated using the  library. The plot shows the data from the txt file in a line graph.